requirements.txt
fastapi>=0.110
uvicorn[standard]
pyyaml
pydantic>=2.0
langgraph>=0.2
langchain-anthropic>=0.3
langchain-core>=0.3
apscheduler>=3.10
# Tool dependencies
httpx>=0.27
tavily-python>=0.3
apify-client>=1.7
pymupdf>=1.24
python-docx>=1.1
